+++ src/app/[locale]/(client)/layout.tsx
@@ -1,44 +1,23 @@
 import Footer from "../../../components/Footer/Footer";
 import Header from "../../../components/Header/Header";
 import "../../globals.scss";
-import type { Metadata } from "next";
-import { notFound } from "next/navigation";
-import { NextIntlClientProvider } from "next-intl";
+// import type { Metadata } from "next";
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-async function getMessages(locale: string) {
-  try {
-    return (await import(`../../../../localization/${locale}.json`)).default;
-  } catch (error) {
-    notFound();
-  }
-}
-
-export async function generateStaticParams() {
-  return ["uk", "en"].map((locale) => ({ locale }));
-}
+// export const metadata: Metadata = {
+//   title: "Create Next App",
+//   description: "Generated by create next app",
+// };
 
 export default async function LocaleLayout({
   children,
-  params: { locale },
 }: {
   children: React.ReactNode;
-  params: { locale: string };
 }) {
-  const messages = await getMessages(locale);
   return (
-    <html lang={locale}>
-      <body>
-        <NextIntlClientProvider locale={locale} messages={messages}>
-          <Header />
-          {children}
-          <Footer />
-        </NextIntlClientProvider>
-      </body>
-    </html>
+    <>
+      <Header />
+      {children}
+      <Footer />
+    </>
   );
 }
