+++ src/app/layout.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+type Props = {
+  children: ReactNode;
+};
+
+export default function RootLayout({ children }: Props) {
+  return children;
+}
