+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
-import Footer from "../../components/Footer/Footer";
-import Header from "../../components/Header/Header";
+import Footer from "../components/Footer/Footer";
+import Header from "../components/Header/Header";
 
 export default function NotFoundPage() {
   return (
